Move the VulkanGraphicsContext to its correct location in Common/GPU/Vulkan
