+++ Common/Common.vcxproj.filters
@@ -789,6 +789,9 @@
     <ClInclude Include="GPU\GraphicsContext.h">
       <Filter>GPU</Filter>
     </ClInclude>
+    <ClInclude Include="GPU\Vulkan\VulkanGraphicsContext.h">
+      <Filter>GPU\Vulkan</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="ABI.cpp" />
@@ -1456,6 +1459,9 @@
     </ClCompile>
     <ClCompile Include="..\ext\pugixml\pugixml.cpp">
       <Filter>ext\pugixml</Filter>
+    </ClCompile>
+    <ClCompile Include="GPU\Vulkan\VulkanGraphicsContext.cpp">
+      <Filter>GPU\Vulkan</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
